% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dnamatch.R
\name{dnamatch}
\alias{dnamatch}
\title{dnamatch}
\usage{
dnamatch(fn, freqfile, reffold = NULL, TAptrn, SIDpat, BIDpat, Thist = 1,
  Tnew = 1, threshLR = 1000, threshHeight = 10, threshStutt = 0.1,
  threshMaj = 0.6, pD = 0, pC = 0, sameKL = FALSE, N = 5000)
}
\arguments{
\item{fn}{A folder with stain files. Full directory must be given.}

\item{freqfile}{A file containing population frequencies for alleles. Full directory must be given.}

\item{reffold}{A folder with stored personal references. Default is no references. Full directory must be given.}

\item{TAptrn}{Filename structure of samples to evaluate. For instance TAptrn="TA-".}

\item{SIDpat}{Pattern of name which is sample ID (SID_BID_CID)=("-S0001_BES00001-14_2014234231"). Here SIDpat="-S". Can also be a vector.}

\item{BIDpat}{Pattern of name which is unique for all samples to evaluate. (SID_BID_CID)=("-S0001_BES00001-14_2014234231"). Here BIDpat="_BES".}

\item{Thist}{Number of days back in time for file to be imported(stains). Samples older than Tnew are declared as OLD.}

\item{Tnew}{Number of days back in time which are defined as NEW files. If today is 2014-11-16, Tnew=3 means that files from 2014-11-13 are declared as NEW.}

\item{threshLR}{Threshold for extracting a match to file.}

\item{threshHeight}{Acceptable peak height (rfu) in stains.}

\item{threshStutt}{Acceptable stutter ratio in stains (relative peak heights) .}

\item{threshMaj}{If second largest allele has ratio (relative to the largest allele) above this threshold, then second allele is part of the major profile (used for extracting major from mixture). If the relative peak height between second and third largest allele has ratio greater than this threshold, no major is assigned.}

\item{pD}{Assumed drop-out rate per marker (parameter in LR model).}

\item{pC}{Assumed drop-in rate per marker (parameter in LR model).}

\item{sameKL}{Boolean whether matches within same case ID number are shown.}

\item{N}{Database size used for estimating population frequencies. Used to assign new allele frequenceis.}
}
\description{
dnamatch is a function for doing contamination search between trace samples (mixtures) and between trace samples and reference samples where the trace samples are selected within a range of time.
}
\details{
dnamatch imports case-stains from genemapper-format automatically and feeds it into a structure for doing comparison matching.

Note: Samplenames must have this format: "SID_BID_CID". The samplename must be unique and consist of SID=Sample ID, BID=Unique ID for all samples, CID=Case ID.

Timestamp="YY-MM-DD-HH-MM-SS" is used as identification key of when the dnamatch function was run.
Match results are stored as a table in matchfile.csv with the timestamp. Matches which are earlier found in matchfile.csv will not be stored again. The column "Checked" can be used for comments.
More details about a given dnamatch run are stored in the folder "session" with corresponding timestamp.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

